+++ src/App.css
@@ -36,3 +36,75 @@
     transform: rotate(360deg);
   }
 }
+
+
+/*NAVIGATION BAR*/
+
+.topnav {
+  background-color: #ff6600;
+  overflow: hidden;
+}
+
+/* Style the links inside the navigation bar */
+.topnav a {
+  float: left;
+  color: #f2f2f2;
+  text-align: center;
+  padding: 50px 50px;
+  text-decoration: none;
+  font-size: 28px;
+}
+
+/* Change the color of links on hover */
+.topnav a:hover {
+  background-color: #ddd;
+  color: black;
+}
+
+/* UF logo */
+.topnav a.active {
+  background-color: #0033ff;
+  color: white;
+}
+
+
+/* DISCLAIMER TEXT*/
+.page_disclaim {
+  background-color: #0033ff;
+  float: right;
+  text-align: center;
+  color: white;
+  padding: 10px 10px;
+  font-size: 19.6px;
+}
+
+.header{
+  padding: 10px;
+  text-align: center;
+  background: #0033ff;
+  color: white;
+  font-size: 19.6px;
+}
+
+/* end of page nav*/
+
+.students {
+  column-count: 1;
+  column-gap: 40px;
+  column-rule-style: solid;
+}
+
+
+.column {
+  float: left;
+  width: 100%;
+  padding: 10px;
+  height: 300px; /* Should be removed. Only for demonstration */
+}
+
+/* Clear floats after the columns */
+.row:after {
+  content: "";
+  display: table;
+  clear: both;
+}
